backend/requirements.txt
fastapi
uvicorn # for running locally
python-dotenv
pinecone-client
openai
redis
pydantic
sentence-transformers
